+++ src/st-breadcrumbs/index.ts
@@ -1,0 +1,1 @@
+export { StBreadcrumbsModule } from './st-breadcrumbs.module';