+++ src/st-breadcrumbs/st-breadcrumbs.module.ts
@@ -1,0 +1,13 @@
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { StBreadCrumbs } from './st-breadcrumbs.component';
+
+
+@NgModule({
+   imports: [CommonModule],
+   exports: [StBreadCrumbs],
+   declarations: [StBreadCrumbs],
+   entryComponents: [StBreadCrumbs],
+   providers: []
+})
+export class StBreadcrumbsModule {}